fix: fix error in command thia2

+++ src/controllers/callback_query/removeKey.js
@@ -1,6 +1,8 @@
 import checkSetAccount from "../../util/checkSetAccount.js";
 import typingMessage from "../../util/tyingMessage.js";
 import Key from "../../model/Key.js";
+import checkPermisson from "../../util/checkPermisson.js";
+
 async function removeKey({ data, message }) {
   // const timeStartSkip = new Date();
   // console.log(data);
@@ -20,6 +22,10 @@ async function removeKey({ data, message }) {
       chat_id,
       message: `Đợi chút để mình xóa Key: *${json.key}*`,
     });
+    if (!checkPermisson(chat_id)) {
+      await editMessage("Bạn không có quyền sử dụng chức năng này");
+      return;
+    }
     await Key.deleteOne({
       key: json.key,
     });
